+++ news.md
@@ -1,6 +1,8 @@
 # tidySEM 0.2.4
 
 * Fixed bug in plot_profiles() when covariances are free
+* Improved starting values for mixture models with free covariances (check and 
+  fix non-positive definite vcov matrices)
 * Documentation incorrectly referred to data_mixture_ordinal, instead of
   data_mix_ordinal
 * Added legend to plot_probs()