+++ news.md
@@ -1,3 +1,9 @@
+# tidySEM 0.2.5
+
+* Add reference to best practices paper, <DOI:10.1080/10705511.2023.2250920>
+* Include vignettes corresponding to the appendices of that paper
+* Correct implementation of `lr_lmr()` using `nonnest2::vuongtest()`
+
 # tidySEM 0.2.4
 
 * Update maintainer email address